// src/app/core/models/ViewModel/units-vm.ts
export class UnitsVM{
    id: any;
    code!: string;
    unitNameAr!: string;
    unitNameEn!: string;
    purposeType!: string;
    officeId!: bigint;
    unitStatus!: number;
    buildingId!: bigint;
    ownerId!: bigint;
    readyForSell!: boolean;
    sellMeterPrice: any;
    sellAreaSize: any;
    sellPrice: any;
    rentMeterPrice: any;
    rentAreaSize: any;
    rentPrice: any;
    isFeatured!: boolean;
    hasPriceOffer!: boolean;
    hasGarage!: boolean;
    garageArea: any;
    vehiclesCount: any;
    customGarageCount: any;
    taxRatio: any;
    taxAmount: any;
    insurranceRatio: any;
    insurranceAmount: any;
        
    }