// src/app/core/models/units.ts
export class Unit {
    id: any;
    code!: string;
    unitNameAr!: string;
    unitNameEn!: string;
    purposeType!: string;
    officeId!: bigint;
    unitStatus!: number;
    buildingId!: bigint;
    floorId!: bigint;
    unitTypeId!:bigint;
    ownerId!: bigint;
    readyForSell!: boolean;
    sellMeterPrice: any;
    sellAreaSize: any;
    sellPrice: any;
    rentMeterPrice: any;
    rentAreaSize: any;
    rentPrice: any;
    isFeatured!: boolean;
    hasPriceOffer!: boolean;
    hasGarage!: boolean;
    garageArea: any;
    vehiclesCount: any;
    customGarageCount: any;
    taxRatio: any;
    taxAmount: any;
    insurranceRatio: any;
    insurranceAmount: any;
    offerPrice:any;
    offerDescription:any;


}